Split out discovery into new file
