+++ homeassistant/components/mqtt/device_tracker/schema_discovery.py
@@ -1,17 +1,16 @@
-"""Support for tracking MQTT enabled devices."""
+"""Support for tracking MQTT enabled devices identified through discovery."""
 import logging
 
 import voluptuous as vol
 
 from homeassistant.components import device_tracker, mqtt
-from homeassistant.components.device_tracker import PLATFORM_SCHEMA, SOURCE_TYPES
+from homeassistant.components.device_tracker import SOURCE_TYPES
 from homeassistant.components.device_tracker.config_entry import TrackerEntity
 from homeassistant.const import (
     ATTR_GPS_ACCURACY,
     ATTR_LATITUDE,
     ATTR_LONGITUDE,
     CONF_DEVICE,
-    CONF_DEVICES,
     CONF_ICON,
     CONF_NAME,
     CONF_UNIQUE_ID,
@@ -23,16 +22,16 @@ from homeassistant.core import callback
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 
-from . import (
+from .. import (
     MqttAttributes,
     MqttAvailability,
     MqttDiscoveryUpdate,
     MqttEntityDeviceInfo,
     subscription,
 )
-from .const import ATTR_DISCOVERY_HASH, CONF_QOS, CONF_STATE_TOPIC
-from .debug_info import log_messages
-from .discovery import MQTT_DISCOVERY_NEW, clear_discovery_hash
+from ..const import ATTR_DISCOVERY_HASH, CONF_QOS, CONF_STATE_TOPIC
+from ..debug_info import log_messages
+from ..discovery import MQTT_DISCOVERY_NEW, clear_discovery_hash
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -40,15 +39,7 @@ CONF_PAYLOAD_HOME = "payload_home"
 CONF_PAYLOAD_NOT_HOME = "payload_not_home"
 CONF_SOURCE_TYPE = "source_type"
 
-PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend(mqtt.SCHEMA_BASE).extend(
-    {
-        vol.Required(CONF_DEVICES): {cv.string: mqtt.valid_subscribe_topic},
-        vol.Optional(CONF_PAYLOAD_HOME, default=STATE_HOME): cv.string,
-        vol.Optional(CONF_PAYLOAD_NOT_HOME, default=STATE_NOT_HOME): cv.string,
-        vol.Optional(CONF_SOURCE_TYPE): vol.In(SOURCE_TYPES),
-    }
-)
-DISCOVERY_PLATFORM_SCHEMA = (
+PLATFORM_SCHEMA_DISCOVERY = (
     mqtt.MQTT_RO_PLATFORM_SCHEMA.extend(
         {
             vol.Optional(CONF_DEVICE): mqtt.MQTT_ENTITY_DEVICE_INFO_SCHEMA,
@@ -65,14 +56,14 @@ DISCOVERY_PLATFORM_SCHEMA = (
 )
 
 
-async def async_setup_entry(hass, config_entry, async_add_entities):
+async def async_setup_entry_from_discovery(hass, config_entry, async_add_entities):
     """Set up MQTT device tracker dynamically through MQTT discovery."""
 
     async def async_discover(discovery_payload):
         """Discover and add an MQTT device tracker."""
         discovery_data = discovery_payload.discovery_data
         try:
-            config = DISCOVERY_PLATFORM_SCHEMA(discovery_payload)
+            config = PLATFORM_SCHEMA_DISCOVERY(discovery_payload)
             await _async_setup_entity(
                 hass, config, async_add_entities, config_entry, discovery_data
             )
@@ -125,7 +116,7 @@ class MqttDeviceTracker(
 
     async def discovery_update(self, discovery_payload):
         """Handle updated discovery message."""
-        config = DISCOVERY_PLATFORM_SCHEMA(discovery_payload)
+        config = PLATFORM_SCHEMA_DISCOVERY(discovery_payload)
         self._setup_from_config(config)
         await self.attributes_discovery_update(config)
         await self.availability_discovery_update(config)
@@ -236,34 +227,3 @@ class MqttDeviceTracker(
     def source_type(self):
         """Return the source type, eg gps or router, of the device."""
         return self._config.get(CONF_SOURCE_TYPE)
-
-
-async def async_setup_scanner(hass, config, async_see, discovery_info=None):
-    """Set up the MQTT tracker."""
-    devices = config[CONF_DEVICES]
-    qos = config[CONF_QOS]
-    payload_home = config[CONF_PAYLOAD_HOME]
-    payload_not_home = config[CONF_PAYLOAD_NOT_HOME]
-    source_type = config.get(CONF_SOURCE_TYPE)
-
-    for dev_id, topic in devices.items():
-
-        @callback
-        def async_message_received(msg, dev_id=dev_id):
-            """Handle received MQTT message."""
-            if msg.payload == payload_home:
-                location_name = STATE_HOME
-            elif msg.payload == payload_not_home:
-                location_name = STATE_NOT_HOME
-            else:
-                location_name = msg.payload
-
-            see_args = {"dev_id": dev_id, "location_name": location_name}
-            if source_type:
-                see_args["source_type"] = source_type
-
-            hass.async_create_task(async_see(**see_args))
-
-        await mqtt.async_subscribe(hass, topic, async_message_received, qos)
-
-    return True
